+++ src/deepboard/gui/components/__init__.py
@@ -10,3 +10,5 @@ from .quick_filter import QuickFilter
 from .plotly2fh import plotly2fasthtml_download
 from .autocomplete import AutoCompleteInput
 from .stat_cell import StatCell
+from .artefact_header import ArtefactHeader, ArtefactFilterOptions
+from .artefactFilter import ArtefactFilter, ArtefactValue